refactor: split test_tools_ci.py (1009 lines) into three focused files
Move check_repo_merge_state tests to new test_tools_ci_merge_state.py and
auto_trigger tests to test_tools_ci_watch.py; reduce test_tools_ci.py to
wait_for_ci/get_ci_status/wait_for_merge_queue tests only. No behavioral changes.
